refactor(worker): extract child-worker supervisor into supervisor.rs (sc-1636)
Move the process-supervision cluster out of lib.rs into supervisor.rs:
the WorkerSpec/SupervisedChild types and the supervise/restart/stop/
terminate/spawn loop plus the auto- and utility-worker spec builders.

Same pattern: `use super::*;` (DiscoveredGpu stays at the root, shared
with gpu.rs; gpu worker-id helpers resolve via the root glob),
pub(crate) on the items reached from run()/tests, glob re-import at the
crate root, worker tests.rs imports repointed to super::supervisor.
WorkerSpec/SupervisedChild fields are pub(crate) since tests inspect
them. Behavior-preserving. lib.rs 4,631 -> 4,414 lines.
