+++ crates/sceneworks-worker/src/supervisor.rs
@@ -1,0 +1,225 @@
+use super::*;
+
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub(crate) struct WorkerSpec {
+    pub(crate) worker_id: String,
+    pub(crate) gpu_id: String,
+}
+
+pub(crate) struct SupervisedChild {
+    pub(crate) spec: WorkerSpec,
+    pub(crate) process: Child,
+    pub(crate) restart_attempt: u32,
+}
+
+pub(crate) async fn supervise_auto_workers(settings: Settings) -> WorkerResult<()> {
+    let gpus = discover_gpus().await;
+    if gpus.is_empty() {
+        let specs = utility_worker_specs(&settings.worker_id, settings.utility_workers);
+        return supervise_children(settings, specs).await;
+    }
+
+    let specs = auto_worker_specs(&settings.worker_id, &gpus);
+    supervise_children(settings, specs).await
+}
+
+/// Spawn the given child workers and keep them running, restarting any that exit
+/// (with backoff) until a shutdown signal arrives.
+pub(crate) async fn supervise_children(
+    settings: Settings,
+    specs: Vec<WorkerSpec>,
+) -> WorkerResult<()> {
+    let mut children = HashMap::new();
+    for spec in specs {
+        let process = start_child_worker(&settings, &spec)?;
+        children.insert(
+            spec.worker_id.clone(),
+            SupervisedChild {
+                spec,
+                process,
+                restart_attempt: 0,
+            },
+        );
+    }
+
+    let mut interval = tokio::time::interval(Duration::from_secs(1));
+    interval.set_missed_tick_behavior(MissedTickBehavior::Delay);
+    loop {
+        tokio::select! {
+            _ = shutdown_signal() => {
+                stop_children(&settings, &mut children).await;
+                return Ok(());
+            }
+            _ = interval.tick() => {
+                restart_exited_children(&settings, &mut children).await?;
+            }
+        }
+    }
+}
+
+pub(crate) async fn restart_exited_children(
+    settings: &Settings,
+    children: &mut HashMap<String, SupervisedChild>,
+) -> WorkerResult<()> {
+    restart_exited_children_with_spawner(settings, children, start_child_worker).await
+}
+
+pub(crate) async fn restart_exited_children_with_spawner<F>(
+    settings: &Settings,
+    children: &mut HashMap<String, SupervisedChild>,
+    mut spawner: F,
+) -> WorkerResult<()>
+where
+    F: FnMut(&Settings, &WorkerSpec) -> WorkerResult<Child>,
+{
+    let mut exited = Vec::new();
+    for (worker_id, child) in children.iter_mut() {
+        if let Some(status) = child.process.try_wait()? {
+            // Advance the attempt once here so the logged ETA and the actual
+            // backoff below both read the same stored value.
+            child.restart_attempt = child.restart_attempt.saturating_add(1);
+            let delay = retry_delay(settings.poll_seconds, child.restart_attempt);
+            emit_json(json!({
+                "event": "worker_exited",
+                "workerId": worker_id,
+                "gpuId": child.spec.gpu_id,
+                "exitCode": status.code(),
+                "restartInSeconds": delay,
+                "reportedAt": now_rfc3339(),
+            }));
+            exited.push(worker_id.clone());
+        }
+    }
+    for worker_id in exited {
+        let Some(mut child) = children.remove(&worker_id) else {
+            continue;
+        };
+        let delay = retry_delay(settings.poll_seconds, child.restart_attempt);
+        tokio::select! {
+            _ = tokio::time::sleep(Duration::from_secs(delay)) => {}
+            _ = shutdown_signal() => {
+                children.insert(worker_id, child);
+                stop_children(settings, children).await;
+                return Ok(());
+            }
+        }
+        let process = spawner(settings, &child.spec)?;
+        child.process = process;
+        children.insert(child.spec.worker_id.clone(), child);
+    }
+    Ok(())
+}
+
+pub(crate) async fn stop_children(
+    settings: &Settings,
+    children: &mut HashMap<String, SupervisedChild>,
+) {
+    for child in children.values_mut() {
+        terminate_child(&mut child.process).await;
+    }
+    let deadline = tokio::time::sleep(Duration::from_secs(
+        settings.shutdown_timeout_seconds.max(1),
+    ));
+    tokio::pin!(deadline);
+    loop {
+        let mut remaining = 0_usize;
+        for child in children.values_mut() {
+            match child.process.try_wait() {
+                Ok(Some(_)) => {}
+                Ok(None) => remaining += 1,
+                Err(_) => {}
+            }
+        }
+        if remaining == 0 {
+            children.clear();
+            return;
+        }
+        tokio::select! {
+            _ = &mut deadline => break,
+            _ = tokio::time::sleep(Duration::from_millis(100)) => {}
+        }
+    }
+    for child in children.values_mut() {
+        let _ = child.process.start_kill();
+        let _ = child.process.wait().await;
+    }
+    children.clear();
+}
+
+pub(crate) async fn terminate_child(child: &mut Child) {
+    #[cfg(unix)]
+    {
+        if let Some(pid) = child.id() {
+            let _ = nix::sys::signal::kill(
+                nix::unistd::Pid::from_raw(pid as i32),
+                nix::sys::signal::Signal::SIGTERM,
+            );
+            return;
+        }
+    }
+    let _ = child.start_kill();
+}
+
+pub(crate) fn start_child_worker(_settings: &Settings, spec: &WorkerSpec) -> WorkerResult<Child> {
+    let executable = std::env::current_exe()?;
+    emit_json(json!({
+        "event": "starting_worker",
+        "workerId": spec.worker_id,
+        "gpuId": spec.gpu_id,
+        "reportedAt": now_rfc3339(),
+    }));
+    let mut command = Command::new(executable);
+    command.envs(child_environment(spec));
+    command.spawn().map_err(Into::into)
+}
+
+pub(crate) fn child_environment(spec: &WorkerSpec) -> BTreeMap<String, String> {
+    let mut env = BTreeMap::new();
+    env.insert("SCENEWORKS_WORKER_CHILD".to_owned(), "1".to_owned());
+    env.insert("SCENEWORKS_WORKER_ID".to_owned(), spec.worker_id.clone());
+    env.insert("SCENEWORKS_GPU_ID".to_owned(), spec.gpu_id.clone());
+    if spec.gpu_id == "cpu" {
+        env.insert("CUDA_VISIBLE_DEVICES".to_owned(), String::new());
+        env.insert("SCENEWORKS_UTILITY_JOBS".to_owned(), "1".to_owned());
+    } else {
+        env.insert("CUDA_VISIBLE_DEVICES".to_owned(), spec.gpu_id.clone());
+        env.insert("SCENEWORKS_UTILITY_JOBS".to_owned(), "0".to_owned());
+    }
+    env
+}
+
+pub(crate) fn auto_worker_specs(base_worker_id: &str, gpus: &[DiscoveredGpu]) -> Vec<WorkerSpec> {
+    let mut specs = gpus
+        .iter()
+        .map(|gpu| WorkerSpec {
+            worker_id: gpu_worker_id(base_worker_id, &gpu.id),
+            gpu_id: gpu.id.clone(),
+        })
+        .collect::<Vec<_>>();
+    specs.push(WorkerSpec {
+        worker_id: cpu_worker_id(base_worker_id),
+        gpu_id: "cpu".to_owned(),
+    });
+    specs
+}
+
+/// Specs for the dedicated CPU/utility worker pool. The first worker keeps the
+/// historical `<base>-cpu` id (so a single-worker setup is unchanged); each
+/// additional worker is suffixed `-1`, `-2`, ... A count of 0 still yields one.
+pub(crate) fn utility_worker_specs(base_worker_id: &str, count: usize) -> Vec<WorkerSpec> {
+    (0..count.max(1))
+        .map(|index| WorkerSpec {
+            worker_id: utility_worker_id(base_worker_id, index),
+            gpu_id: "cpu".to_owned(),
+        })
+        .collect()
+}
+
+pub(crate) fn utility_worker_id(base_worker_id: &str, index: usize) -> String {
+    let cpu_id = cpu_worker_id(base_worker_id);
+    if index == 0 {
+        cpu_id
+    } else {
+        format!("{cpu_id}-{index}")
+    }
+}
